requirements.txt
Flask
Flask-API
flake8
paho-mqtt
pytest
requests
